requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
httpx>=0.25.0
google-generativeai>=0.4.0
scikit-learn>=1.3.0
numpy>=1.26.0
pandas>=2.1.0
reportlab>=4.0.0
pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.26.0
python-multipart>=0.0.6
